requirements.txt
airflow-provider-kafka
apache-airflow-providers-microsoft-azure==4.1.0
apache-airflow-providers-slack==6.0.0
azure-storage-blob
astronomer-providers